Share the browser initialization code.
Adds logging of when we fail to create the browser widget.
