+++ gapic/src/main/com/google/gapid/views/Licenses.java
@@ -15,6 +15,7 @@
  */
 package com.google.gapid.views;
 
+import static com.google.gapid.widgets.Widgets.withSizeHints;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.logging.Level.SEVERE;
 
@@ -25,18 +26,14 @@ import com.google.gapid.proto.service.Service.ClientAction;
 import com.google.gapid.util.Messages;
 import com.google.gapid.widgets.DialogBase;
 import com.google.gapid.widgets.Theme;
+import com.google.gapid.widgets.Widgets;
 
 import org.eclipse.jface.dialogs.IDialogConstants;
 import org.eclipse.swt.SWT;
-import org.eclipse.swt.SWTError;
-import org.eclipse.swt.browser.Browser;
-import org.eclipse.swt.browser.LocationAdapter;
-import org.eclipse.swt.browser.LocationEvent;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Shell;
-import org.eclipse.swt.widgets.Text;
 
 import java.io.IOException;
 import java.util.logging.Logger;
@@ -45,7 +42,7 @@ import java.util.logging.Logger;
  * Dialog to display the licenses used by the application in an HTML browser window.
  */
 public class Licenses {
-  private static final Logger LOG = Logger.getLogger(Licenses.class.getName());
+  protected static final Logger LOG = Logger.getLogger(Licenses.class.getName());
 
   private Licenses() {
   }
@@ -61,33 +58,9 @@ public class Licenses {
       @Override
       protected Control createDialogArea(Composite parent) {
         Composite area = (Composite)super.createDialogArea(parent);
-
-        Browser browser;
-        try {
-          browser = new Browser(area, SWT.NONE);
-        } catch (SWTError e) {
-          // Failed to initialize the browser. Show it as a plain text widget.
-          Text text = new Text(
-              area, SWT.MULTI | SWT.READ_ONLY | SWT.BORDER | SWT.H_SCROLL | SWT.V_SCROLL);
-          text.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true));
-          text.setText(readLicenses(false));
-          return text;
-        }
-
-        GridData data = new GridData(SWT.FILL, SWT.FILL, true, true);
-        data.widthHint = 600;
-        data.heightHint = 400;
-        browser.setLayoutData(data);
-        browser.setText(readLicenses(true));
-        browser.addLocationListener(new LocationAdapter() {
-          @Override
-          public void changing(LocationEvent event) {
-            if ("about:blank".equals(event.location)) {
-              browser.setText(readLicenses(true));
-            }
-          }
-        });
-
+        Control browser = Widgets.createBrowser(area, readLicenses());
+        browser.setLayoutData(
+            withSizeHints(new GridData(SWT.FILL, SWT.FILL, true, true), 700, 768));
         return area;
       }
 
@@ -98,17 +71,11 @@ public class Licenses {
     }.open();
   }
 
-  protected static String readLicenses(boolean html) {
+  protected static String readLicenses() {
     try {
       String result = Resources.toString(Resources.getResource("text/licenses.html"), UTF_8);
-      if (html) {
-        // Linkify links.
-        result = result.replaceAll(
-            "(https?://[^\\s]+(?:[^\\s.,;:!?'\"()\\[\\]{}]))", "<a href='$1'>$1</a>");
-      } else {
-        // De-HTML the text.
-        result = result.replaceAll("<[^>]+>", "");
-      }
+      result = result.replaceAll(
+          "(https?://[^\\s]+(?:[^\\s.,;:!?'\"()\\[\\]{}]))", "<a href='$1'>$1</a>");
       return result;
     } catch (IOException | IllegalArgumentException e) {
       LOG.log(SEVERE, "Failed to load the licenses", e);
